+++ src/resources/api/index.ts
@@ -1,10 +1,4 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 export { API } from './api';
-export {
-  V1,
-  type V1ListProjectsResponse,
-  type V1UploadFileResponse,
-  type V1ListProjectsParams,
-  type V1UploadFileParams,
-} from './v1/index';
+export { V1 } from './v1/index';