+++ src/resources/top-level.ts
@@ -1,6 +1,6 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
-export type HealthCheckResponse = unknown;
+export type HealthCheckResponse = string;
 
 export declare namespace TopLevel {
   export { type HealthCheckResponse as HealthCheckResponse };