+++ CHANGELOG.md
@@ -1,5 +1,44 @@
 # Changelog
 
+## 0.6.0 (2026-06-17)
+
+Full Changelog: [v0.5.1...v0.6.0](https://github.com/cellect-ai/cellect-sdk-typescript/compare/v0.5.1...v0.6.0)
+
+### Features
+
+* **api:** api update ([cde67b7](https://github.com/cellect-ai/cellect-sdk-typescript/commit/cde67b773ffb8f0bfbbe4f9c8b1128943c9abfe8))
+* **api:** api update ([781f457](https://github.com/cellect-ai/cellect-sdk-typescript/commit/781f457036e4fc4a652ef56e6c3328f088ccb227))
+* **api:** api update ([82f6076](https://github.com/cellect-ai/cellect-sdk-typescript/commit/82f6076b94c25572eaa2563245f9c5ac5653c77c))
+* **api:** api update ([b7fff95](https://github.com/cellect-ai/cellect-sdk-typescript/commit/b7fff953fd95b7eac3bd4dc66e03275d72c41ea6))
+* **api:** api update ([4067b7c](https://github.com/cellect-ai/cellect-sdk-typescript/commit/4067b7c0939b4bb7a059beec8b8191e09834500a))
+* **api:** api update ([b174932](https://github.com/cellect-ai/cellect-sdk-typescript/commit/b17493206f60d5695e9cd854fa599bca364006f4))
+* **api:** api update ([8fd1bfa](https://github.com/cellect-ai/cellect-sdk-typescript/commit/8fd1bfad57f19b56df40fce9180ab27d89c1720f))
+* **api:** api update ([e967281](https://github.com/cellect-ai/cellect-sdk-typescript/commit/e967281ed2d02db2b026f4310370f57384975fb2))
+* **api:** api update ([59128ff](https://github.com/cellect-ai/cellect-sdk-typescript/commit/59128ff5dc51affde22e4ac095f2e32b367be1c8))
+* **api:** api update ([3c5a7a4](https://github.com/cellect-ai/cellect-sdk-typescript/commit/3c5a7a42a499822fedbe21ada1f3d21e00812115))
+
+
+### Bug Fixes
+
+* **client:** send content-type header for requests with an omitted optional body ([d57d73c](https://github.com/cellect-ai/cellect-sdk-typescript/commit/d57d73c622532d30f0969f95eb6c96e005883457))
+* **mcp:** correct code tool API endpoint ([9fa362a](https://github.com/cellect-ai/cellect-sdk-typescript/commit/9fa362aae2abad5322defd32486faf4f06a07096))
+* **mcp:** return correct lines on typescript errors ([23e2f70](https://github.com/cellect-ai/cellect-sdk-typescript/commit/23e2f7011714dfb2fb94d423b93e871de730255c))
+
+
+### Chores
+
+* **client:** fix logger property type ([eb49922](https://github.com/cellect-ai/cellect-sdk-typescript/commit/eb4992225925f780cd7b9120e440d151f08b7880))
+* **internal:** codegen related update ([af44d7c](https://github.com/cellect-ai/cellect-sdk-typescript/commit/af44d7ccec5b7f5e204658e323be2bfa7bf0b3ee))
+* **internal:** codegen related update ([0ed4218](https://github.com/cellect-ai/cellect-sdk-typescript/commit/0ed42188d8ccc854e520311ac209c6c0411cf8e0))
+* **internal:** codegen related update ([b5a6625](https://github.com/cellect-ai/cellect-sdk-typescript/commit/b5a662551b450c588998025a1bc52bf47bdea2d1))
+* **internal:** codegen related update ([2b95478](https://github.com/cellect-ai/cellect-sdk-typescript/commit/2b95478ba82c1a203a4ca4584e1a1ab7c6617512))
+* **internal:** codegen related update ([b84fcf9](https://github.com/cellect-ai/cellect-sdk-typescript/commit/b84fcf98b7aef8220c98a68746cb20ca3ee8468d))
+* **internal:** codegen related update ([8fe24b6](https://github.com/cellect-ai/cellect-sdk-typescript/commit/8fe24b663e71225d55426089c952dbc223851598))
+* **internal:** codegen related update ([5f9d29c](https://github.com/cellect-ai/cellect-sdk-typescript/commit/5f9d29ce0304bcc101e4ebbba2052b2e2be2940f))
+* **internal:** update `actions/checkout` version ([b523ec2](https://github.com/cellect-ai/cellect-sdk-typescript/commit/b523ec2194f79dbd1f3c57aa8bbea5398ebaacb2))
+* **internal:** upgrade babel, qs, js-yaml ([1a93005](https://github.com/cellect-ai/cellect-sdk-typescript/commit/1a93005eaddd19d40dc34de33488b834d7665580))
+* **internal:** upgrade eslint ([e64064a](https://github.com/cellect-ai/cellect-sdk-typescript/commit/e64064a0a764f0cd865eca3e9bc2fe76338f671e))
+
 ## 0.5.1 (2025-10-23)
 
 Full Changelog: [v0.5.0...v0.5.1](https://github.com/cellect-ai/cellect-sdk-typescript/compare/v0.5.0...v0.5.1)